Differential gene expression, donor + treatment
To determine the DGE profiles (for mouse separately; for human and HBV genes separately) relative to uninfected controls, of self-assembling co-cultures of primary human hepatocytes (SACC-PHHs) (co-cultured with 3T3J mouse non-parenchymal cells) mono- infected with HBV or co-infected with HBV/HDV at 8 and 28 days post-infection. Here, donor and treatment are factors in the design.